// client/src/ducks/invoice/actions-types.ts
export enum ActionType {
    INVOICE_LIST_REQUEST = "INVOICE_LIST_REQUEST",
    INVOICE_LIST_SUCCESS = "INVOICE_LIST_SUCCESS",
    INVOICE_LIST_FAILURE = "INVOICE_LIST_FAILURE",
    INVOICE_CREATE_REQUEST = "INVOICE_CREATE_REQUEST",
    INVOICE_CREATE_SUCCESS = "INVOICE_CREATE_SUCCESS",
    INVOICE_CREATE_FAILURE = "INVOICE_CREATE_FAILURE",
    INVOICE_DELETE_REQUEST = "INVOICE_DELETE_REQUEST",
    INVOICE_DELETE_SUCCESS = "INVOICE_DELETE_SUCCESS",
    INVOICE_DELETE_FAILURE = "INVOICE_DELETE_FAILURE",
    INVOICE_EDIT_REQUEST = "INVOICE_EDIT_REQUEST",
    INVOICE_EDIT_SUCCESS = "INVOICE_EDIT_SUCCESS",
    INVOICE_EDIT_FAILURE = "INVOICE_EDIT_FAILURE",
    INVOICE_PAY_SUCCESS = "INVOICE_PAY_SUCCESS",
    INVOICE_PAY_REQUEST = "INVOICE_PAY_REQUEST",
    INVOICE_PAY_FAILURE = "INVOICE_PAY_FAILURE",
}